requirements.txt
pip==25.0.1
PyYAML==6.0.2
setuptools==75.8.2
toml==0.10.2
wheel==0.45.1
